Combined the line-edit and vertex edit tools
