+++ bokehjs/src/coffee/models/tools/draw/vertex_edit_tool.ts
@@ -24,9 +24,9 @@ export class VertexEditToolView extends DrawToolView {
     const renderers = this._select_event(e, false, this.model.renderers);
 
     // If we did not hit an existing line, clear node CDS
-    const point_renderer = this.model.point_renderer;
-    const point_ds = point_renderer.data_source;
-    const point_glyph = point_renderer.glyph;
+    const vertex_renderer = this.model.vertex_renderer;
+    const point_ds = vertex_renderer.data_source;
+    const point_glyph = vertex_renderer.glyph;
     let [pxkey, pykey] = Object.getPrototypeOf(point_glyph)._coords[0];
     [pxkey, pykey] = [point_glyph.attributes[pxkey].field, point_glyph.attributes[pykey].field];
     if (!renderers.length) {
@@ -66,7 +66,7 @@ export class VertexEditToolView extends DrawToolView {
   }
 
   _tap(e: BkEv): void {
-    const renderer = this.model.point_renderer;
+    const renderer = this.model.vertex_renderer;
     const point = this._map_drag(e.bokeh.sx, e.bokeh.sy, renderer);
     if (point == null) {
       return;
@@ -75,6 +75,7 @@ export class VertexEditToolView extends DrawToolView {
     const append = e.srcEvent.shiftKey != null ? e.srcEvent.shiftKey : false;
     const indices = ds.selected['1d'].indices.slice(0);
     this._select_event(e, append, [renderer]);
+    this._select_event(e, append, this.model.renderers);
     if (ds.selected['1d'].indices.length) {
       return;
     }
@@ -94,25 +95,57 @@ export class VertexEditToolView extends DrawToolView {
 
   _pan_start(e: BkEv): void {
     const append = e.srcEvent.shiftKey != null ? e.srcEvent.shiftKey : false;
-    this._select_event(e, append, [this.model.point_renderer]);
+    this._select_event(e, append, [this.model.vertex_renderer]);
+    this._select_event(e, false, this.model.renderers);
     this._basepoint = [e.bokeh.sx, e.bokeh.sy];
   }
 
   _pan(e: BkEv): void {
-    const renderer = this.model.point_renderer;
-    const ds = renderer.data_source;
-    const point = this._map_drag(e.bokeh.sx, e.bokeh.sy, renderer);
-    const basepoint = this._map_drag(this._basepoint[0], this._basepoint[1], renderer);
-    if (!ds.selected['1d'].indices.length || point == null || basepoint == null || this._selected_renderer == null) {
+    const [bx, by] = this._basepoint;
+    if (this._selected_renderer == null) {
+      for (const renderer of this.model.renderers) {
+        const basepoint = this._map_drag(bx, by, renderer);
+        const point = this._map_drag(e.bokeh.sx, e.bokeh.sy, renderer);
+        if (point == null || basepoint == null) {
+          continue;
+        }
+
+        const ds = renderer.data_source;
+        const glyph = renderer.glyph;
+        let [xkey, ykey] = Object.getPrototypeOf(glyph)._coords[0];
+        [xkey, ykey] = [glyph.attributes[xkey].field, glyph.attributes[ykey].field];
+        const [x, y] = point;
+        const [px, py] = basepoint;
+        const [dx, dy] = [x-px, y-py];
+        for (const index of ds.selected['1d'].indices) {
+          const xs = ds.data[xkey][index];
+          const ys = ds.data[ykey][index];
+          for (let i = 0; i < ys.length; i++) {
+            xs[i] += dx;
+            ys[i] += dy;
+          }
+        }
+        ds.change.emit(undefined);
+      }
+      this._basepoint = [e.bokeh.sx, e.bokeh.sy];
+	  return;
+	}
+
+    const vertex_renderer = this.model.vertex_renderer;
+    const ds = vertex_renderer.data_source;
+    const point = this._map_drag(e.bokeh.sx, e.bokeh.sy, vertex_renderer);
+    const basepoint = this._map_drag(bx, by, vertex_renderer);
+    if (!ds.selected['1d'].indices.length || point == null || basepoint == null) {
       return;
     }
 
     // If a Point is selected drag it
+    const glyph = vertex_renderer.glyph;
     const [x, y] = point;
     const [px, py] = basepoint;
     const [dx, dy] = [x-px, y-py];
-    let [xkey, ykey] = Object.getPrototypeOf(renderer.glyph)._coords[0];
-    [xkey, ykey] = [renderer.glyph.attributes[xkey].field, renderer.glyph.attributes[ykey].field];
+    let [xkey, ykey] = Object.getPrototypeOf(glyph)._coords[0];
+    [xkey, ykey] = [glyph.attributes[xkey].field, glyph.attributes[ykey].field];
     for (const index of ds.selected['1d'].indices) {
       ds.data[xkey][index] += dx;
       ds.data[ykey][index] += dy;
@@ -123,21 +156,32 @@ export class VertexEditToolView extends DrawToolView {
   }
 
   _pan_end(_e: BkEv): void {
-    this.model.point_renderer.data_source.selected['1d'].indices = [];
+    this.model.vertex_renderer.data_source.selected['1d'].indices = [];
     if (this._selected_renderer) {
       this._selected_renderer.data_source.properties.data.change.emit(undefined);
-    }
+    } else {
+	  for (const renderer of this.model.renderers) {
+        renderer.data_source.selected['1d'].indices = [];
+        renderer.data_source.properties.data.change.emit(undefined);
+      }
+	}
     this._basepoint = null;
   }
 
   _keyup(e: BkEv): void {
     if ((e.keyCode === 8) && this.model.active) {
-      this._delete_selected(this.model.point_renderer);
+      if (this._selected_renderer) {
+        this._delete_selected(this.model.vertex_renderer);
+      } else {
+        for (const renderer of this.model.renderers) {
+          this._delete_selected(renderer);
+        }
+      }
     }
   }
 
   deactivate(): void {
-    const renderer = this.model.point_renderer;
+    const renderer = this.model.vertex_renderer;
     const ds = renderer.data_source;
     const glyph = renderer.glyph;
     let [pxkey, pykey] = Object.getPrototypeOf(glyph)._coords[0];
@@ -151,7 +195,8 @@ export class VertexEditToolView extends DrawToolView {
 }
 
 export class VertexEditTool extends DrawTool {
-  point_renderer: GlyphRenderer
+  vertex_renderer: GlyphRenderer
+  drag: boolean
 
   tool_name = "Vertex Edit Tool"
   icon = "bk-tool-icon-vertex-edit"
@@ -164,5 +209,6 @@ VertexEditTool.prototype.type = "VertexEditTool"
 VertexEditTool.prototype.default_view = VertexEditToolView
 
 VertexEditTool.define({
-  point_renderer: [ p.Instance ],
+  drag: p.Bool,
+  vertex_renderer: [ p.Instance ],
 })
