+++ bokehjs/src/coffee/models/tools/index.ts
@@ -9,7 +9,6 @@ export {ZoomOutTool}       from "./actions/zoom_out_tool"
 export {ButtonTool}        from "./button_tool"
 export {DrawTool}          from "./draw/draw_tool"
 export {BoxDrawTool}       from "./draw/box_draw_tool"
-export {LineEditTool}      from "./draw/line_edit_tool"
 export {PointDrawTool}     from "./draw/point_draw_tool"
 export {PolyDrawTool}      from "./draw/poly_draw_tool"
 export {VertexEditTool}    from "./draw/vertex_edit_tool"
